+++ NEWS.md
@@ -6,7 +6,7 @@
  if the data is close to the right hand axis. You can stop this by setting a manual x-limit
  for the right hand side. Vertically divided time series graphs always add a margin. 
  ([#292](https://github.com/angusmoore/arphit/pull/292))
-
+ * Export a list of graphs as a GIF animation ([#293](https://github.com/angusmoore/arphit/pull/293))
 ## Breaking changes
 
 ## Bugfixes